requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
psycopg2-binary
djangorestframework
django-filter
